Backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-generativeai==0.8.4
python-dotenv==1.0.0
gunicorn==21.2.0
bcrypt==4.0.1